requirements.txt
flask==3.1.2
pymongo==4.15.1
python-dotenv==1.1.1
requests==2.32.5
mcp==1.0.0
